Refactor code to a modular approach
