+++ cli/const.go
@@ -1,0 +1,12 @@
+package cli
+
+// VM launcher environment variables
+const (
+	VMLauncherWorkdir = "$HOME/vms"
+	SSHPublicKeyFile  = "$HOME/.ssh/id_rsa.pub"
+)
+
+// Container Images
+const (
+	VMLauncherContainerImage = "vmgo/vmgo"
+)
