+++ cli/ssh.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"bufio"
